[FIX] helpdesk_stock: remove suitable_picking_ids for returns w/o ticket_id
Before this commit:
`suitable_picking_ids` is populated with all `done` pickings when
returning a transfer without a `ticket_id`, which crashes the frontend
upon loading.

Steps to reproduce:
    1. Install Stock & Helpdesk
    2. Create thousands of transfers and mark them as done
    3. Return a transfer

After this commit:
Set `suitable_picking_ids` to False when the transfer doesn't have
a `ticket_id`, as the fields are not used in that case anyway.

opw-3232319

closes odoo/enterprise#38595

X-original-commit: e009cf2cc1363f3d7bc9417c214357772fb255db
Signed-off-by: William Henrotin (whe) <whe@odoo.com>