[FIX] project: fix spacing issue in project task kanban view
Before this commit, in the project module, project task kanban view is extended
in the project_enterprise module. In that module, planned_date_begin field is
added to create space in project kanban view.

This commit ensures that if the task has a starting and ending date
field then it will be shown to view.

task-3231708

closes odoo/enterprise#39503

X-original-commit: 6f77e2c7673be73f4aadd2f3a52f981476e89e32
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>