[IMP] social: prevent the read time out due to connectivity issues
While loading the feed page with function( _fetch_page_posts)
in social marketing module, which is facing connectivity issue
due to loss of network or slow internet the page is getting
read time out.
This is happening when the page is being refreshed/synchronized.
To avoid showing an error to the end user, we ignore those read
timeouts.As refreshing the streams/statistics is a recurring work,
they will get refreshed again later.

sentry id-3953488576

closes odoo/enterprise#38466

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>