[FIX] knowledge: stop the emoji insults
This commit fixes a bug introduced in c33fe997f270a086150b628ab1021e9469e89c02
where the addIcon function was modified to accomodate the Discuss refactor.

Before this commit when adding a random icon to an article, the random
selection of icons would always return the banned emojis, which is the opposite
of the desired behavior.

Now the icon is always an authorized emoji.

task-3239114

closes odoo/enterprise#38433

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>