[IMP] account_accountant: bank recon enhancements
Links:
The journal dashboard has two links to the bank recon: the journal name, and the reconcile button.
Both lead to the same place but don't match in terms of filters and domains.
This commits removes the domain filter and applies a search_default instead - aligning to the reconcile button.

Table Layout:
Prevent the horizontal scroll on the Bank Recons lines when the partner or label are too long. Use Ellipses instead.

AML Filters:
Combine payable and receivable filters in Customer/Vendor
Add a filter for miscellaneous journal entries

View Partner:
Allow to open the partner form from the Manual Operations tab

Rename Statement Line Filters:
Received -> Incoming
Paid -> Outgoing

Remove the tutorial and enhance the noContentHelper screens with rainbow man, and buttons:
Before this PR, the tutorial was displayed when there were no transactions visible.
Now, the noContentHelper screen is displayed instead. It includes buttons to clear the filters if there are any, and to go back to the previous action or dashboard. When transactions have been reconciled, rainbowman is displayed instead.

Task-3221418

closes odoo/enterprise#39017

Signed-off-by: Laurent Smet <las@odoo.com>