[FIX] documents_sign: Disable adding signed documents tags
Steps:
- Install Sign and Documents modules
- Go to Sign, upload a pdf template and save
- Modify the template, Show "template properties"
- Try to add a Signed document tag and save

Issue:
- The form view doesn't close automatically (But the call to the server is done)
- When closing the form the created tag doesn't appear

Cause:
Same to this commit: https://github.com/odoo/enterprise/commit/5458c14f8863b632183dba3b0eea4f045269433e
Problem with OWL2 which isn't not yet resolved.

Beside that, the PO ensured that it's not functionnally correct to be able to create tags from that form view.
So for now we will just disable the create and edit option for the ``documents_tags_ids`` field.

opw-3227339

closes odoo/enterprise#42347

X-original-commit: 08fd1cabc89574d27693446bf652a62cb8a8f366
Signed-off-by: Djoumatchoua Eteil Junior (etdj) <etdj@odoo.com>