[IMP] {web,timesheet}_grid: avoid re-rendering grid renderer when hover a cell
Before this commit, when the user hovers a cell, the grid renderer is
re-renderered because the highlighted cells are done with a state
and each time the state is changed, the grid renderer is re-rendered.
Also, the GridCell component is mounted when the cell is mounted and
so it is mounted when the grid renderer is re-rendered. Because of
this system of highlighted cells, the grid view in `All Timesheets`
could lagged when the user scrolls the view.

This commit avoid re-rendering the grid renderer when the user hovers
a cell by dynamically adding a class (`o_grid_highlighted`) to notice
the cell is highlighted.
This class is added on all cells in the same row and columns of the cell
hovered. Another class is added (`o_grid_row_highlighted`) on all cells
in the same rows to add some others styles for those cells.
A last class is added `o_grid_cell_highlighted` on the cell hovered to
add a specific style.

part of task-3226579
task-3283082

X-original-commit: 4a03df67247e70e89edf9a676cc2ec1a657b7265
Part-of: odoo/enterprise#40460