[FIX] web_studio: invisible modifier value
To reproduce
============
- on any app that has a notebook
- on studio make one of the pages invisible
- try to make this page visible again (impossible)

Problem
=======
When setting the page invisible the checkbox is not checked, so clicking on it again will trigger action
to make the page invisible.
The issue comes from the fact that the `checked` value of this checkbox is binded to `widget.state.attrs.invisible`
but from 16, the invisible modifier is not included in `attrs` directly but in `attrs.modifiers`.

Solution
========
use `widget.state.attrs.modifiers.invisible`

opw-3233079

closes odoo/enterprise#38489

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>