[IMP] stock_barcode: split both the operations from barcode commands
Currently, all operation From barcode command are in single report.

In this commit,I have Split the storage locations from barcode commands and now we
are printing reports separately. Also change limits the number of locations and
picking type sent to the pdf generation because currenlty limits are 100 so in
this case every 4th page print only 16 barocdes where as we have to print 28 barcode
in on every page.

TaskID - 2579195

closes odoo/enterprise#38901

Related: odoo/odoo#116844
Signed-off-by: Steve Van Essche <svs@odoo.com>