[FIX] l10n_de_reports: make datev visible for DACH countries
Some customers in Austria are exporting their data using the
DateV export available in `l10n_de_reports`. We can show the
button for all DACH countries (countries where DateV is active)

In the future, we shoud get this logic out of the German module,
and make sure that it works for all of these countries.

closes odoo/enterprise#39741

X-original-commit: 9c0f4b799084ef37a33dec67db493846aa16837e
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>