[IMP] hr_contract_sign : Display signatures sent to all employee emails
Right now the signature request smartbutton only displays sign requests sent to work emails, but an employee might not have a work email set yet when using the recruitment app for example. So the behavior after this PR is : take work email if it exists, if not take the private email, if no email is set notify the user.
task - 3240123

closes odoo/enterprise#38535

Signed-off-by: Kevin Baptiste <kba@odoo.com>