[FIX] industry_fsm_report: template empty name
To reproduce the issue:
1. Create a field service task
2. Create a customer
3. When creating the customer, go to contacts & addresses,
and add a blank delivery address
4. Add the customer to the task
5. Start then stop the timer
6. Send report by clicking on action > send report
7. Error message: "Failed to render inline_template template"

Error: The report should have been generated

When generating the pdf name for the worksheet, the condition
was testing object.parter_id to know if the record was set,
to know if it could use the name.
However, in this instance, the name can be left empty and
the record partner would still be created, which would
cause the testing of partner_id to be true, while name
is empty, and cause an error when trying to concatenate.

The condition has been adapted to test if the name was set
to be able to use it to name the worksheet.

OPW-3210956

closes odoo/enterprise#38852

X-original-commit: 50d36db865edfad0673372a2d47da3e31e2cf873
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Demany Antoine (ande) <ande@odoo.com>