[REF] l10n_mx_edi: Reword Mexican EDI independant of account_edi
- Remove the dependency to account_edi:
Since the rework of the send & print wizard, the MX EDI is now triggered from the
send & print wizard and no longer at the post like what account_edi does.

- Improve the tracking of CFDI documents from the invoices:
Since the government want to now exactly what part of each invoices is currently paid
and the part of taxes paid for each invoice, it was complicated to navigate from one
screen to another to send payments to the government.
Now, we add a full tracking of payments directly on invoices to improve the tracking of
CFDI documents.

- Integrate stock:
The Delivery Guide are now using the same engine as invoices/payments to track the CFDI documents.

- Prepare for global invoice:
We will need the CFDI on POS order. To prepare the model for this integration, this commit adds a new
model l10n_mx_edi.document allowing to track CFDI documents for any model.
Also, the invoice itself is no longer used as a record in the cfdi template. This will ease the integration
of the global invoice since we need a way to have a CFDI file for multiple invoices.

closes odoo/enterprise#44052

Task: 3069324
Related: odoo/odoo#128395
Related: odoo/upgrade#4942
Signed-off-by: Laurent Smet (las) <las@odoo.com>