[IMP] sale_subscription: do not rely on payment.transaction callback
In subscription, we have two ways to do actions once the payment is validated (state done/authorized):

- payment callback, synchronous.
- reconcile after done, synchronous

The aim of the task is to avoid using the callback to use reconcile after done.
callback are only used in subscription and are subject to issues if something
crashes, we can be in a weird state in the cron with a draft invoice, a good payment and a confirmed transaction.

This commit also introduce a pending payment flag on order to allow bypass orders in the cron while a payment
is being processed.

Taskid: 3258616
Part-of: odoo/enterprise#39195