[FIX] knoweldge: name_create receives correct input
This commit fixes a bug where in the kanban view of all the articles, when you
group them by their Parent Article and try to add a new column the view would
crash.

This crash is caused by the name_create function when called by the rpc service
on the client's side. The function is unable to receive the necessary input in
order for it to work properly, leading to a crash.

To fix this function we added the decorator `api.model` above it, this way when
we the function via the rpc service, it will receive the correct parameters and
work properly.

task-3244062

closes odoo/enterprise#38533

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>