[FIX] industry_fsm: not copy fsm_is_sent field
When the "Worksheets" parameter is activated, it is possible to send reports.
Once a report has been sent, the button no longer appears (desired behavior).

If we duplicate the task, the button still does not appear.

In order to be able to review the sent report button on a duplicate task
(condition in the `_compute_display_send_report_buttons` method),
the `fsm_is_sent` field must not be copied.

opw-3226588

closes odoo/enterprise#38813

X-original-commit: 045696edc267c351bd2cef43de43de1c44a2beec
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>