[IMP] appointment: use different description fo...
r exported calendar events

We now generate a different summary and description for ics files
and google calendar exportation.

This is a lot more clear for the people booking the appointment.
We also don't expect users to get an ics file via this method
so always giving the organizer details makes sense.

task-3232952

closes odoo/enterprise#38570

Related: odoo/odoo#116211
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>