[FIX] website_sale_renting: be able to unarchive "List View Variants"
Current behaviour:
Broken xpath when unarchiving one of the "List View Variants" (check
ticket for more info).

Expected behaviour:
There should be no broken xpath.

Steps to reproduce:
- Install website_sale_renting
- Make sure Variants are active
- In debug mode, edit the views, (with active or inactive filter)
  search for the views "List View of Variants".
- If you archive the one with the external ID, you cannot unarchive
  it after that.

Reason for the problem:
Depending on the order of installation of the views, the `t-placeholder`
may not be a direct child of the element that has the class
`js_main_product`.

Fix:
Correct the XPath so that `t-placeholder` is not necessarilly a child of
the element with the class `js_main_product`.

Affected versions:
- 16.0
- saas-16.1
- saas-16.2
- master

opw-3236843

closes odoo/enterprise#38577

Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>