[IMP] stock_barcode: confirm location
When a product is scanned and this product is reserved in another
location than the current one, the application will ask the user to
confirm they want to take this product in the current location by
scanning the source's barcode.

The purpose of this commit is to avoid scan error by taking a product in
the wrong location because user could forget to scan the location
between two different products. Imagine the following case:
- There is a delivery for Product A (reserved in Shelf 1) and Product B
  (reserved in Shelf 2);
- First, the user scans Shelf 1, than scans Product A. So far so good;
- Now the user scans Product B. The question is: do they want to take
  Product B from Shelf 1, or did they forget to scan Shelf 2 first ?

This issue is relevant only when the products are reserved and the scan
of the source location is mandatory.

To handle this case, a product reserved in a location can be taken from
another location only if this other location was scanned just before the
product. A notification informs the user they have to scan the location.
So, if we go back to our previous example, the user has to scan again
Shelf 1 just before to scan Product B if they really wants to take it
from there.

task-3007755

closes odoo/enterprise#41502

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>