[IMP] account_reports: account_codes engine: improve query performance
On large databases, the query previously ran by the account_codes engine was becoming slower, even when computing a single line, with a single expression, referencing a single account code in its formula. This was due to the fact we were limiting the computation to only the move lines done on the accounts we wanted: filtering them took time. With this commit, we don't do this filtering anymore, and compute more things than we need; but this makes the computation faster (balance sheet opens ~3 times faster on odoo.com).

closes odoo/enterprise#38675

X-original-commit: 9e381b6bcb5707df296c8a72ce68d054095f2e7d
Signed-off-by: John Laterre (jol) <jol@odoo.com>