[FIX] l10n_nl_intrastat: exported csv inconsistent with report
Reproduction:
1. Install Accounting, l10n_nl_accounting, l10n_nl_intrastat, go to the
setting of accounting and check customer addresses
2. Switch to the nl company, create an invoice with a customer in
Ireland and choose the Delivery Address in Italy.
3. Go to Reporting->Intrastat Report, the country code of this invoice
is Italy, which is correct
4. Click EXPORT (CBS), check the exported csv file, the country code
column is IE instead of IT

Reason: we use the invoice’s partner address for “Dispatch” type of
invoices instead of the delivery address. Dispatch means we are the
seller and we send the product to the clients

Fix: change the country_dest_code as the delivery address

opw-3079132

closes odoo/enterprise#38727

X-original-commit: 04c68026ff940e38130d11fe50fbd162dd602a53
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Liu Jinjiu (jili) <jili@odoo.com>