[FIX] appointment: use latest stop and longer lifespan
The appointment invitation links have two issues:
first, before this fix, we used the latest meeting created
and checked that it was not ending more than three
months ago, which seems arbitrary and random.

If it was, then we deleted the whole appointment
invite record in the garbage collector. Now we make
sure that we take the latest stop value of linked
meetings.

Also, (Internal need) 3 months is too little in practice
as many salespeople see their link expire in signature.
To keep the number of links under control, we still
keep a limit, but increase it to 6 months.

A test is added to ensure that we only remove invitations
(older than 6 months old) having their latest meeting ending
older than 6 months old. This is done in a new file for
tests concerning the appointment invite model.

Task-3241362

closes odoo/enterprise#39077

X-original-commit: 436dabbe7a6474494c3d2c2c12bab3d0f349b730
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>