[FIX] knowledge: padding readonly articles
This commit fixes a styling bug that was corrected in 24aad7b666a398ea9e78f4dbf26eba9715f47332
where readonly articles would have more padding than editable ones.

Since the OWLification of the knowledge icon to a field widget, when the field
is empty the class `o_field_empty` is added to the element and this class some
height to it.
This style would be taken into account when the article would be readonly.
To fix this we added a css rule that sets a min-height of 0px when the icon
field is empty, removing the unnecessary paddding inside knowledge.

We also removed a css rule that would remove the padding left for the control
panel inside an embedded view when not in `full_width`.

task-3245853

closes odoo/enterprise#39135

X-original-commit: 6c10f4eb4ca782e0678ef7097418fb7e035959b1
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>