[FIX] web_studio: Properly add measure field ids to studio arch
In the pivot view editor sidepanel, the "Measures" pivot view field did
not show the currently selected measures shown in the view and crashed
when trying to select a new one.

This was caused by the "studio_pivot_measure_field_ids" attribute not
being set on the studio view arch.

*This issue also prevented easy access to the measure field ids for the
web_studio owl refactor*

closes odoo/enterprise#38635

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>