[IMP] l10n_pe_edi: raise error directly
The aim of this commit is to improve the reliability by throwing the
error not found directly instead of waiting for the government to tell
the customer Odoo didn't make the job correctly.

Before this commit:
If the national bank account isn't present the edi system continues its
jobs and doesn't fill the PayeeFinancialAccount node in the xml.
We expect then the customer to find out there is something wrong with it
by having the governement.

After this commit:
If the national bank account will be needed to generate the EDI, Odoo
throws a UserError asking the user to setup the national bank account.

closes odoo/enterprise#38636

Task-id: None
Signed-off-by: Josse Colpaert <jco@odoo.com>