[REF] spreadsheet_*: move model and callbacks to action
For the sharing task (3045808), a share button will be added to the
control panel. That means it will need to have the model.

Currently, it's instanciated deep in <SpreadsheetComponent> and therefore
not available to the action component.

This commit moves the model creation up to the action component.

It also moves lots of other stuff from <SpreadsheetComponent> to
the action component which were implemented by passing props callbacks to
<SpreadsheetComponent>, basically just to be called without anything
specific from <SpreadsheetComponent>.
Moving everything up in the action is just much simpler.

closes odoo/enterprise#38640

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>