[ADD] l10n_br_avatax*: add Brazilian tax calculation via Avatax
Add tax computation for Brazil. Just like in the US, taxes in Brazil
are highly variable and depend on all sorts of details of the seller,
customer and the products being sold. This module provides these
details to Avatax and Avatax tells us what taxes we should charge.

Depending on account_avatax was considered and attempted but resulted
in complex code and features that weren't needed, e.g.
avatax_tax_date, product.avatax.category and the concept of committing
records.

On top of that some changes needed to be made anyway to the core of
account_avatax, most importantly having to support price-included
taxes and subtracted taxes (negative "price-excluded" taxes).

Another major difference is the need to have separate sandbox and
production credentials and the fact that this module requires IAP:
customer accounts have to be created under Odoo Inc's "master" Avatax
account. A separate IAP service will create these "sub-accounts" and
proxy the tax calculation requests and perform an IAP charge for each
of them.

A final difference is that negative lines are not allowed by the
Brazilian government, because of this the module will spread these
amounts over the positive lines (similar to l10n_mx_edi). There's also
the concept of "transport" costs which are distributed over lines
based on the subtotal.

The downside of not inheriting account_avatax is duplicated code (~32%
more, from 1054 LOC (excl mocked API responses) to 1388 LOC). The goal
is to extract the duplicated code to a common module in master.

Community PR: odoo/odoo#128216
IAP PR: odoo/iap-apps#610

task-2968770

Alternative approach to #39970

closes odoo/enterprise#43797

Signed-off-by: Josse Colpaert <jco@odoo.com>