[FIX] planning:  Fix grey separator visible in the front end view
Description of the issue/feature this PR addresses:
On the website when the planning is open, a noticeable grey separator can be
seen below it.

Current behavior before PR:
When any button is not visible on the footer then there is an empty grey
separator visible

Desired behavior after PR is merged:
When the buttons are not visible then no separator should be displayed

Fix:
To ensure that the buttons are only visible when intended, the display property
should be set to none when they are not meant to be displayed and to block
when they are meant to be displayed.

task - 3229105

closes odoo/enterprise#45323

X-original-commit: 289513485e646529801d7e7eed27c86b48bcc041
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>