[FIX] account_online_synchronization: handle redirect errors
With the B2B task, the way we handle error is different than before,
we want to be able to redirect the user directly to what he should
do when needed. Since the redirect error can happen in different
method call, use a custom error for this and also fix an issue
when we need to link a new institution.
Also handle a use case where the redirect happened while connecting
with an existing account.

task-3247332

closes odoo/enterprise#39887

X-original-commit: 12ba46eb1008bc0cd070b719701a1430be08f9f2
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>