[FIX] timesheet_grid: adapt hovered for grid button timer component
Before this commit, the highlighted cells have been changed to be able
to mount the GridCell component without re-rendering the grid renderer
component  when a grid cell with value is hovered.

This commit adapts the code when the user hovers a grid timer button
component to have the same behavior as before.

part of task-3226579
task-3283082

X-original-commit: 6a57f09aafa9e4464b2f01bc59d272f1e7fccfa4
Part-of: odoo/enterprise#40460