[FIX] mrp_workorder: avoid writing when reading QCP
Reading a QCP may trigger a useless write request on it.

To reproduce the issue:
(Use demo data)
1. Install `quality_mrp_workorder`
2. Enable debug mode
3. Open an existing Quality Control Point
4. Debug > View Metadata

Error: the _Latest Modification Date_ is defined to now, but nothing
did modify the QCP

When opening the QCP, it leads to the compute method of
`component_ids`. Here is the issue: in this method, we also write on
the field `component_id`, which is a stored one. On ORM-side, it
will realize that the value is already defined to `False` and will
not write it again, but the writing process will already be started,
so an update of the last user and last modification date will still
be sent to the database. This point could be improved on ORM side
but, still, we should avoid writing something to a field while
computing another one (especially during a read request).

Initially, the reset of `component_id` has been added to an onchange,
so that we remove its value if the user changes the test type (see [1])

[1] ed37271dac04fe0d532e2888ced1f457fd35a2a0

OPW-3147450

closes odoo/enterprise#39055

X-original-commit: e72494f3992ef7d0b0f5ddad12ff2a597b86e7b6
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>