[FIX] timesheet_grid_holidays: display user error if timer on time off
Before this commit, when the user tries to start a timer on time off
task, a traceback is occured because the timer on the timesheet is not
started since the timesheet is in time off task.

This commit fixes the issue by displaying a user error when the user
tries to start a timer on timesheet in a time off task.

task-3245969

X-original-commit: 4de004ed6a3a023daede1666162b715c1422d19d
Part-of: odoo/enterprise#38910