[FIX] sale_subcription: payment_exception copy=False
Currently the `payment_exception` field has no parameter copy=False.
When a existing subscription order with "payment failure" status gets
duplicated, `payment_exception = true` gest carried over, which will lead
to a UserError when the customer tries to create an invoice from the
duplicated record.
The root is this filter call
https://github.com/odoo/enterprise/blob/16.0/sale_subscription/models/sale_order.py/#L1140-L1141

Before this change:
Duplicating a subscription with `payment_exception = true`, copies the
status to the duplicate and blocks any invoice to be created out of the
duplicated record.

After this change:
Duplicating a subscription with 'payment failure' should correctly keep
the `payment_exception = false` and function as expected, i.e. invoices
can be created.

opw-3216747

closes odoo/enterprise#38788

Signed-off-by: Arnaud Joset <arj@odoo.com>