[FIX] analytic_enterprise: fix traceback in year range
Before this commit, when the user wants to see the grid view in year
range, a traceback is occured because the fetch method in JS is
miswritten to get the period to display in the grid view.

This commit fixes the issue to get the grid view with the period
according to the fiscal year set on the current company.

task-3245969

X-original-commit: e99d78174b2b97a5a076b7efc03530d11b6ad181
Part-of: odoo/enterprise#38910