[IMP] mail_enterprise,web_enterprise,*: Support Web Push Notification
* = test_mail_enterprise

WebPush allows sending data to the user's browser/app(PWA) even when
tab/app is closed. Web push is a "constant" link between the
ServiceWorker of browser/app and a WebPush server.

Note that each browser has its own custom WebPush server.
e.g.:

* Chrome: https://fcm.googleapis.com/
* Firefox: https://updates.push.services.mozilla.com/
* Safari: https://web.push.apple.com/
* Edge: https://wns2-ln2p.notify.windows.com/

WebPush introduces some cryptographic notion to ensure some the
reliability of the data sent:

1. VAPID: "Voluntary Application Server Identification" is the standard
   used to generate the public and the private keys to sign the message
   between the browser and WebPush server
2. JWT: "JSON Web Token" is the standard used to sign the payload to the
   WebPush server
3. ECE: "Encrypted Content-Encoding" is the standard used by WebPush to
   encrypt the data of the payload to avoid sending RAW data
   outside trusted network.

Simplified steps how WebPush works:

1. The Javascript code of a web page subscribes to the WebPush server
   (using the VAPID key generated at mail_entreprise install).
2. The WebPush server replay with a subscription (and some other infos
   like the unique URL endpoint per subscription where to send a
   notification)
3. The application (odoo-bin in our case) sends a post request to the
   WebPush server using the specific URL endpoint of the user (using JWT
   and ECE).
4. The WebPush server sends back to the browser the encrypted payload.
5. The browser decrypts the payload and sends it to the ServiceWorker
   linked to the subscription.

Here a [Sequence diagram](https://w3c.github.io/push-api/index.html#fig-example-flow-of-events-for-subscription-push-message-delivery-and-unsubscription) of all interactions to process a web push notification.

In Odoo, we use WebPush to send Notification to the user.

This commit aims to have a parity with the Android/iOS Mobile App at
the notification level.

Notes:

There are some ways to encrypt (ECE) the message for WebPush:
* AESGCM128: this is a draft
* AESGCM: very well documented
* AES128GCM: RFC8188 Standard encoding

We implement only the RFC one as it is the only one implemented in all
major updated browsers (Chrome, Firefox, Safari, Edge, ...)

You need to allow the desktop "Notification" and "Push" inside your
browser. For iOS Devices, it only works on iOS 16.4+ and it's requiring
Odoo to first be added to the Home Screen. It's delivered silently,
meaning no sound, vibration, haptics or screen wake.

Notifications are sent directly if there are less than 5 notifications,
otherwise we use a cron triggered immediately.

Also, we have changed the value of the "QueryCount" as `mail_enterprise`
executes a new query to search the devices associated with the partner.

We have added a "try/except" for any Exception before the
`push_to_end_point` method as we want to avoid blocking a normal flow
just for a not mandatory push notification if something happens during
the push to the endpoint.
See: odoo/enterprise@d0ae70103dd1c169debe8fcb0918f80310c85e42

Links:

- https://www.rfc-editor.org/rfc/rfc8030
- https://www.rfc-editor.org/rfc/rfc8188
- https://www.rfc-editor.org/rfc/rfc8291
- https://www.rfc-editor.org/rfc/rfc8292
- https://w3c.github.io/push-api/index.html
- https://autopush.readthedocs.io/en/latest/http.html
- https://web.dev/push-notifications-web-push-protocol/
- https://github.com/web-push-libs/encrypted-content-encoding
- https://github.com/web-push-libs/pywebpush
- https://github.com/web-push-libs/vapid
- https://caniuse.com/push-api

Task ID: 3123678

closes odoo/enterprise#43577

Related: odoo/odoo#127198
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Co-authored-by: Romeo Fragomeli <rfr@odoo.com>