[IMP] knowledge: allow template and file commands in tables
Authorize `/file` and `/template` commands in a table and update the css to use
the size of the container instead of the size of the viewport.

There was not really a reason to prevent using those commands in a table except
the sizing, but we could let the user decide whether he wants those blocks or
not (works fine with tables of <4 columns).

Update the styling of toolbars for a more precise handling of the positioning of
text and buttons (since `/file` and `/template` commands can now be way smaller
more frequently).

Task-3247197

closes odoo/enterprise#38827

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>