[FIX] web_grid: display `Add a Line` button near Today one when sample data
Before this commit, when sample data is displayed is no possible for
the user to create a timesheet when create inline feature is enabled
in the grid view.

This commit displays the `Add a Line` button near `Today` when
no content helper is displayed and the create action is available.

task-3245969

X-original-commit: 1f845ea4a524cafee42c4b891b59eeabe37f3907
Part-of: odoo/enterprise#38910