[FIX] sale_subscription: is_subscription should not depends on subscription_state
Before this commit, is_subscription depended on subscription_state. It was causing some
issues because wben the subscription_state was updated, fields depending on is_subscription
were recomputed. The subscription_state dependance was only used for the upsell case.
As the upsell values are set at creation and the recurrence_id is readonly, we decide to remove
the dependance to avoid bugs due to recompute.

taskid 3232957

closes odoo/enterprise#38886

X-original-commit: 359a0d24d29ea98c8f8c93dad27047d8426c73ca
Related: odoo/odoo#116820
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>