[FIX] web_grid: use the same list and form defined in current action
Before this commit, when the user clicks on magnifier glass on a grid
cell to see the records aggregated in that cell, the list and form views
could be different than the ones defined in the current action because
the action called is a new action created and the views used are the first
ones fetched. And so, in Accounting > Analytic Items, the mgnifier glass
of a grid cell could load a list view for timesheets.

This commit will search and uses the views of the current action if
the action is a Action Window ("ir.actions.act_window") and only the
list and form views will be used. If one is missing then, the first one
fetched will be fetched.

task-3245969

closes odoo/enterprise#38910

X-original-commit: f256dd41f6b9b594d091600cd12855eaa402076f
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>