[FIX] planning: Fix the recurrence alert display on create
Description of the issue/feature this PR addresses:
In the planning module, When adding a new shift, if you select the repeat field,
the recurrence alert will appear on the header

Current behavior before PR:
When the repeat field is selected, a recurrence alert will be displayed in the
header,

Desired behavior after PR is merged:
At the time of adding shift, no recurrence alert should be displayed. it should
be displayed when editing the shift.

Fix:
When adding a new shift, the system sets the recurrence alert as invisible by
default, but for existing shifts, the alert is visible. This condition ensures
that users are not prompted with unnecessary alerts when creating new shifts

task-3229105

closes odoo/enterprise#39100

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>