[FIX] account_reports: reports on mobile
Fixes an issue with the reports when opening them
on mobile (more specifically, ipad air).
The report width isn't correct at this time.

By adding a computed width in the class, we keep
the current behaviour in desktop/wide screen but
also allow the report to take the whole width in
smaller screen.

Task id # 3249004

closes odoo/enterprise#38856

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>