[FIX] account_reports: reports on mobile
Fixes an issue with the reports when opening them
on mobile (more specifically, ipad air).
The report width isn't correct at this time.

By adding a computed width in the class, we keep
the current behaviour in desktop/wide screen but
also allow the report to take the whole width in
smaller screen.

Task id # 3249004

closes odoo/enterprise#39141

X-original-commit: 7758aeb35e793b992cf627f76a79417dc76aafef
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>