[IMP] l10n_mx_edi_website_sale: display the RFC on the additional step
1. Display the RFC on the additional step while asking for an invoice.
This will remind the eCommerce client that he needs to provide his RFC
number while requiring an Invoice.

2. Display the additional step even if the system parameter "automatic
invoicing" is deactivated. This will allow eCommerce clients to require
and provide the additional information if they require an invoice, even
if the eCommerce is not configured to create invoices automatically.

task-3250896

closes odoo/enterprise#38870

Signed-off-by: Laurent Smet <las@odoo.com>