[FIX] knowledge: apply existing css rules to permission panel
Since the refactoring of the permission panel, the custom css rules of
the permission panel no longer apply. Those rules had the following
effects on the component:

1. It highlighted in blue the current user in the member list.
2. When the user was unable to change the permissions of a user, the
   caret of the selection field was hidden and the cursor of the user
   was turned into a red sign (cursor: not-allowed) on mouse hover to
   indicate that the field can not be edited.
3. The buttons to remove a member from the member list only appeared
   when hovering a user from the list.
4. The font size of the email address was slightly reduced.

To fix the issue, we will update the selectors of the permission panel
component to ensure that the custom css rules will be properly applied
on it as it was the case before.

Related: https://github.com/odoo/enterprise/pull/35261
task-3251343

closes odoo/enterprise#39618

X-original-commit: a28a2c7933717fda6a2d5d2a7dc9c1c13806632f
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>
Signed-off-by: Banken Julien (jbn) <jbn@odoo.com>