[IMP] web_studio, *: use supportedOptions from the field metadata
*: knowledge

This commit removes the need to declare and detail each available option
for a specific widget. Now, this list of supported options are detailed
directly on the field widget's object, with the supportedOptions attribute.

Only the properties that are tied to a type of field remain in a specific
file of web_studio (field_type_properties.js). It is now easier to maintain
and update the available options and attributes that are editable in studio,
since a developer can only document it in the widget implementation file,
without the need to update a list in studio.

task-3259617

closes odoo/enterprise#39847

Related: odoo/odoo#118713
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>